View comunica directamente com Model para obter instrucoes; handle de excepções passa a ser feito no controller; o programa so faz exit se a operação não tiver executado com sucesso
